110 comprobar salidas enpoint

+++ code/CanaryTrailsAPI/src/main/java/es/iespuertodelacruz/mp/canarytrails/controller/v2/RutaControllerV2.java
@@ -89,9 +89,9 @@ public class RutaControllerV2 {
     public ResponseEntity<?> createRuta(@RequestBody RutaEntradaCreateDto dto){
 
         Ruta ruta = rutaMapper.toEntityCreate(dto);
+        ruta.setAprobada(false);
 
-        Usuario usuario = usuarioService.findById(dto.usuario());
-        ruta.setUsuario(usuario);
+        //ruta.setUsuario el creador
 
         for( int id : dto.faunas()){
             Fauna fauna = faunaService.findById(id);
@@ -141,9 +141,9 @@ public class RutaControllerV2 {
         //TODO: comprobar que la ruta es la suya y no está aprobada
 
         Ruta ruta = rutaMapper.toEntityUpdate(dto);
+        ruta.setAprobada(false);
 
-        Usuario usuario = usuarioService.findById(dto.usuario());
-        ruta.setUsuario(usuario);
+        //ruta.setUsuario el creador
 
         if(dto.faunas() != null && !dto.faunas().isEmpty()) {
             for (int id : dto.faunas()) {
@@ -172,16 +172,6 @@ public class RutaControllerV2 {
             }
         }
 
-        if(dto.comentarios() != null && !dto.comentarios().isEmpty()) {
-            for (int id : dto.comentarios()) {
-                Comentario comentario = comentarioService.findById(id);
-                if (comentario != null && !ruta.getComentarios().contains(comentario)) {
-                    comentario.setRuta(ruta);
-                    ruta.getComentarios().add(comentario);
-                }
-            }
-        }
-
         if(dto.municipios() != null && !dto.municipios().isEmpty()) {
             for (int id : dto.municipios()) {
                 Municipio municipio = municipioService.findById(id);
@@ -191,15 +181,11 @@ public class RutaControllerV2 {
             }
         }
 
-        boolean actualizada;
-
         try{
-            actualizada = rutaService.update(ruta);
+            return ResponseEntity.ok(rutaService.update(ruta));
         } catch (RuntimeException e){
             return ResponseEntity.badRequest().body(e.getMessage());
         }
-
-        return ResponseEntity.ok(rutaService.update(ruta));
     }
 
     @PostMapping(value = "/upload/{id}", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
@@ -215,6 +201,10 @@ public class RutaControllerV2 {
             mensaje = "" + namefile;
 
             Ruta ruta = rutaService.findById(id);
+
+            if(ruta == null){
+                return ResponseEntity.notFound().build();
+            }
 
             if(!ruta.getFotos().contains(namefile)){
                 ruta.getFotos().add(namefile);
@@ -241,6 +231,10 @@ public class RutaControllerV2 {
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deleteRuta(@PathVariable("id") int id){
         //TODO: comprobar que la ruta es suya
+
+        if(rutaService.findById(id) == null){
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(rutaService.deleteById(id));
     }
